requirements.txt
tweepy
praw
discord
transformers
scikit-learn
pandas
web3
python-dotenv